move solve_lp.h to include folder. Typedef MT and VT in polytopes' classes. Use exceptions in solve_lp.h.
